+++ tests/test_visitors/test_tokenize/test_primitives/test_numbers/test_number_meaningless_zeros.py
@@ -3,6 +3,7 @@
 import pytest
 
 from wemake_python_styleguide.violations.consistency import (
+    BadNumberSuffixViolation,
     NumberWithMeaninglessZeroViolation,
 )
 from wemake_python_styleguide.visitors.tokenize.primitives import (
@@ -11,12 +12,16 @@ from wemake_python_styleguide.visitors.tokenize.primitives import (
 
 
 @pytest.mark.parametrize('number', [
+    '0.10',
+    '21.5400',
+
     '0x01',
+    '0x0B',
     '0x00A',
 
     '0e00',
     '0e01',
-    '1.5e+010',
+    '1.5e010',
     '1.5e-010',
 
     '0o01',
@@ -24,6 +29,9 @@ from wemake_python_styleguide.visitors.tokenize.primitives import (
 
     '0b0001',
     '0b01',
+
+    '-0.10',
+    '-21.5400',
 
     '-0x01',
     '-0x00A',
@@ -38,12 +46,15 @@ from wemake_python_styleguide.visitors.tokenize.primitives import (
     '-0b0001',
     '-0b01',
 
+    '+0.10',
+    '+21.5400',
+
     '+0x01',
     '+0x00A',
 
     '+0e00',
     '+0e01',
-    '+1.5e+010',
+    '+1.5e010',
     '+1.5e-010',
 
     '-0o01',
@@ -75,13 +86,47 @@ def test_meaningless_zeros(
 
 
 @pytest.mark.parametrize('number', [
+    '0X0A',
+    '0E09',
+    '0B01',
+    '0O07',
+])
+def test_meaningless_zeros_and_case(
+    parse_tokens,
+    assert_errors,
+    default_options,
+    primitives_usages,
+    number,
+    mode,
+):
+    """Ensures that numbers raise two violations."""
+    file_tokens = parse_tokens(mode(primitives_usages.format(number)))
+
+    visitor = WrongNumberTokenVisitor(default_options, file_tokens=file_tokens)
+    visitor.run()
+
+    assert_errors(visitor, [
+        BadNumberSuffixViolation,
+        NumberWithMeaninglessZeroViolation,
+    ])
+
+
+@pytest.mark.parametrize('number', [
+    '-1',
+    '1234567890',
+
+    '0.0',
+    '0.5',
+    '25.05',
+    '10.001',
+
     '0x0',
     '0x10',
     '0xA00',
 
     '0e0',
     '0e10',
-    '1.5e+10',
+    '1.5e10',
     '1.5e-100',
 
     '0o0',
@@ -92,13 +137,18 @@ def test_meaningless_zeros(
     '0b1',
     '0b100000',
 
+    '-0.0',
+    '-0.5',
+    '-25.05',
+    '-10.001',
+
     '-0x0',
     '-0x10',
     '-0xA00',
 
     '-0e0',
     '-0e10',
-    '-1.5e+10',
+    '-1.5e10',
     '-1.5e-100',
 
     '-0o0',
@@ -109,13 +159,18 @@ def test_meaningless_zeros(
     '-0b1',
     '-0b100000',
 
+    '+0.0',
+    '+0.5',
+    '+25.05',
+    '+10.001',
+
     '+0x0',
     '+0x10',
     '+0xA00',
 
     '+0e0',
     '+0e10',
-    '+1.5e+10',
+    '+1.5e10',
     '+1.5e-100',
 
     '+0o0',
@@ -147,18 +202,21 @@ def test_correct_zeros(
     'as_string = "{0}"',
 ])
 @pytest.mark.parametrize('number', [
+    '0.300',
     '0x0FF',
-    '1.5e+01',
+    '1.5e01',
     '0o011',
     '0b01001',
 
+    '-0.300',
     '-0x0FF',
-    '-1.5e+01',
+    '-1.5e01',
     '-0o011',
     '-0b01001',
 
+    '+0.300',
     '+0x0FF',
-    '+1.5e+01',
+    '+1.5e01',
     '+0o011',
     '+0b01001',
 ])