+++ tests/test_visitors/test_tokenize/test_primitives/test_numbers/test_partial_floats.py
@@ -3,6 +3,7 @@
 import pytest
 
 from wemake_python_styleguide.violations.consistency import (
+    NumberWithMeaninglessZeroViolation,
     PartialFloatViolation,
 )
 from wemake_python_styleguide.visitors.tokenize.primitives import (
@@ -58,3 +59,32 @@ def test_correct_float(
     visitor.run()
 
     assert_errors(visitor, [])
+
+
+@pytest.mark.parametrize('primitive', [
+    '.0500',
+    '.00',
+    '+.0500',
+    '+.00',
+    '-.0500',
+    '-.00',
+])
+def test_double_incorrect_float(
+    parse_tokens,
+    assert_errors,
+    assert_error_text,
+    default_options,
+    primitives_usages,
+    primitive,
+    mode,
+):
+    """Ensures that partial floats raise a warning."""
+    file_tokens = parse_tokens(mode(primitives_usages.format(primitive)))
+
+    visitor = WrongNumberTokenVisitor(default_options, file_tokens=file_tokens)
+    visitor.run()
+
+    assert_errors(visitor, [
+        PartialFloatViolation,
+        NumberWithMeaninglessZeroViolation,
+    ])