+++ wemake_python_styleguide/violations/consistency.py
@@ -65,6 +65,7 @@ Summary
    MultilineConditionsViolation
    WrongMethodOrderViolation
    NumberWithMeaninglessZeroViolation
+   PositiveExponentViolation
 
 Consistency checks
 ------------------
@@ -109,6 +110,7 @@ Consistency checks
 .. autoclass:: MultilineConditionsViolation
 .. autoclass:: WrongMethodOrderViolation
 .. autoclass:: NumberWithMeaninglessZeroViolation
+.. autoclass:: PositiveExponentViolation
 
 """
 
@@ -1488,7 +1490,7 @@ class NumberWithMeaninglessZeroViolation(TokenizeViolation):
     Forbids to use meaningless zeros.
 
     We discorauge using meaningless zeros in
-    binary, octal, hex, and expanentional numbers.
+    float, binary, octal, hex, and expanentional numbers.
 
     Reasoning:
         There are ~infinitive ways to write these numbers
@@ -1496,7 +1498,7 @@ class NumberWithMeaninglessZeroViolation(TokenizeViolation):
         ``0b1`` is the same as ``0b01`` and ``0b001``.
         How a language can be called consistent
         if you can write numbers in a infinite ways?
-        It hurts readability and understanding of your programms.
+        It hurts readability and understanding of your code.
 
     Solution:
         Remove meaningless leading zeros.
@@ -1504,10 +1506,10 @@ class NumberWithMeaninglessZeroViolation(TokenizeViolation):
     Example::
 
         # Correct:
-        numbers = [0b1, 0o2, 0x5, 10e10]
+        numbers = [1.5, 0b1, 0o2, 0x5, 10e10]
 
         # Wrong:
-        numbers = [0b00000001, 0o0002, 0x05, 10e010]
+        numbers = [1.50, 0b00000001, 0o0002, 0x05, 10e010]
 
     .. versionadded:: 0.12.0
 
@@ -1515,3 +1517,32 @@ class NumberWithMeaninglessZeroViolation(TokenizeViolation):
 
     error_template = 'Found number with meaningless zeros: {0}'
     code = 339
+
+
+@final
+class PositiveExponentViolation(TokenizeViolation):
+    """
+    Forbids to extra ``+`` signs in the exponent.
+
+    Reasoning:
+        Positive exponent is positive by default,
+        there's no need to write an extra ``+`` sign.
+        We enforce consistency with this rule.
+
+    Solution:
+        Remove meaningless ``+`` sign from the exponent.
+
+    Example::
+
+        # Correct:
+        number = 1e1 + 1e-1
+
+        # Wrong:
+        number = 1e+1
+
+    .. versionadded:: 0.12.0
+
+    """
+
+    error_template = 'Found exponent number with positive exponent: {0}'
+    code = 340